sc-7338 GDS User UI: refact certificate review component
