+++ web/gds-user-ui/src/components/CertificateReview/CertificateReviewLayout.tsx
@@ -1,0 +1,23 @@
+import { Box, Stack } from '@chakra-ui/react';
+import { ReactNode } from 'react';
+
+type CertificateReviewLayoutProps = {
+  children: ReactNode;
+};
+
+function CertificateReviewLayout({ children }: CertificateReviewLayoutProps) {
+  return (
+    <Box
+      border="1px solid #DFE0EB"
+      fontFamily={'Open Sans'}
+      color={'#252733'}
+      bg={'white'}
+      fontSize={18}
+      p={5}
+      px={5}>
+      <Stack>{children}</Stack>
+    </Box>
+  );
+}
+
+export default CertificateReviewLayout;
