+++ web/gds-user-ui/src/utils/getColorScheme.ts
@@ -1,0 +1,9 @@
+const getColorScheme = (status: string | boolean) => {
+  if (status === 'yes' || status === true) {
+    return 'green';
+  } else {
+    return 'orange';
+  }
+};
+
+export default getColorScheme;
